+++ spec/patches/i18n_spec.rb
@@ -6,4 +6,14 @@ RSpec.describe I18n do
     expect(described_class).to receive(:original_locale=).with(described_class.locale)
     described_class.locale = described_class.locale
   end
+
+  it 'transliterate character' do
+    expect(described_class.character_to_group('Ć')).to eq 'C'
+  end
+
+  it 'return character when polish' do
+    with_locale(:pl) do
+      expect(described_class.character_to_group('Ć')).to eq 'Ć'
+    end
+  end
 end